requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
discord-py-slash-command==4.2.1
discord.py==2.0.1
frozenlist==1.3.1
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
joblib==1.2.0
multidict==6.0.2
numpy==1.23.3
oauthlib==3.2.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclickup==0.1.4
pyparsing==3.0.9
python-dotenv==0.21.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
